Formatting corrections
Sometimes, the automatic scaling made the resulting spreadsheet too big,
so that a single row moved to a second page. Apparently, libreoffice uses
an int as the the scale factor and rounds if it gets a float (Arghh).

Also use 5mm reserve on both, width and heigth, previously, we used
5mm for the width and 12mm for the height, and also used a different
reserve when adding vertical stretch to the lines.