Implement new page style for Waagenliste
The list for the scales (and that one alone) is now supposed to
contain the prices for non-members and the origin country as well,
to replace all the small handwritten labels.

That entrails a lot of other changes, because a single A4 sheet will
not offer enough space (except when written too dense), so we print
on 2 A4 Papers in Landscape mode. In that case however, we need to add
column labels that repeat on every printed  page.

- add horCenter and horRight as horizontal alignment constants
- add hcenter and hright as options to ColumnDef
- Sheet now has an optional argument titlerows, default is 0
  This is the number of rows repeated on top of every printed page
- The style of a row is now variable, we have a style with horizontal
  and vertical lines (the previous one) and an additional one
  where only left and right have a vertical line and background
  alternates whited and grey. A different style function is called
  for each style
- A new function to fill the title rows
- function Waagenliste adapted accordingly