Add development script with python controlled libreoffice calc
The script spawns off a libreoffice instance that is listening
on a unix domain socket. It then connects to the instance, initalizes
a few convient variables and finally goes into python interactive mode.

TAB completion show the possible methods and properties on an object,
and you can control the calc instance from python with immediate visual
feedback.