Fix large integer errors
If an object obtained from the database is a large integer (and 'large'
means, not representable with 32 bit, sigh), Python sees it as 'numberic',
but LibreOffice throws a runtime error. Some article numbers are that big
however, so simply limit numeric values to less than 2 * 10^9

Also make the EAN columns for "Lose Ware" wider to show the full number.