Try to really fix the exceptions
The underlying problem is, that the EAN column has type double (in the
database that is). That never is a problem, as long as the value has
less than 9 decimal places, because in that case, the automatic
conversion to integer or string does the right thing,  If tha value
is too big however, and EAN numbers have 13 decimal places which indeed
is too big, we either get integers that are too big for poor LibreOffice,
or floats with a .0 or worse, in scinetific notation.

So we now let the databse deliver the EAN as a string, needing two
casts (sigh). and treat it as String.

Doing so appears to be working now in the fixed layout page as well.