Objectify queries
There was a lot of copy and paste because each query just came with
its own SQL as a mostly constant string, and these had just a few
subtle differences.

This of course calls for a clase, so here it is. A base class Query,
that is designed to be subclassed by just providing it with a few
different class variables. These specify the Fields for the query,
the field's short names as a string and fixed conditions in addition
to the usual LadenID=Plattsalat.

The different selectors "interne Waren Gruppe" (iWG),
"Lieferanden ID" (LieferID), and Warengruppe (WG), can be set as
keyword arguments iwg, liefer and wg respectively in the class
constructor. They can either be a single string or a list of
strings, resulting in = or IN conditions.

The run() method then always performs the query and
returns the result as a list of tuples as before.